combMat {mclust}	R Documentation

Combining Matrix

Description

Create a combining matrix

Usage

combMat(K, l1, l2)

Arguments

K
The original number of classes: the matrix will define a combining from K to (K-1) classes.

l1
Label of one of the two classes to be combined.

l2
Label of the other class to be combined.

Value

If z is a vector (length K) whose kth entry is the probability that an observation belongs to the kth class in a K-classes classification, then combiM %*% z is the vector (length K-1) whose kth entry is the probability that the observation belongs to the kth class in the K-1-classes classification obtained by merging classes l1 and l2 in the initial classification.

Author(s)

J.-P. Baudry, A. E. Raftery, L. Scrucca

See Also

clustCombi, combiPlot

[Package mclust version 5.4.2 Index]