dens {mclust}	R Documentation

Density for Parameterized MVN Mixtures

Description

Computes densities of observations in parameterized MVN mixtures.

Usage

dens(modelName, data, logarithm = FALSE, parameters, warn=NULL, ...)

Arguments

modelName
A character string indicating the model. The help file for mclustModelNames describes the available models.

data
A numeric vector, matrix, or data frame of observations. Categorical variables are not allowed. If a matrix or data frame, rows correspond to observations and columns correspond to variables.

logarithm
A logical value indicating whether or not the logarithm of the component densities should be returned. The default is to return the component densities, obtained from the log component densities by exponentiation.

parameters
The parameters of the model:

pro

The vector of mixing proportions for the components of the mixture.

mean

The mean for each component. If there is more than one component, this is a matrix whose kth column is the mean of the kth component of the mixture model.

variance

A list of variance parameters for the model. The components of this list depend on the model specification. See the help file for mclustVariance for details.

warn
A logical value indicating whether or not a warning should be issued when computations fail. The default is warn=FALSE.

...
Catches unused arguments in indirect or list calls via do.call.

Value

A numeric vector whose ith component is the density of the ith observation in data in the MVN mixture specified by parameters.

See Also

cdens, mclust.options, do.call

Examples

## Not run: 
faithfulModel <- Mclust(faithful)
Dens <- dens(modelName = faithfulModel$modelName, data = faithful,
             parameters = faithfulModel$parameters)
Dens

## alternative call
do.call("dens", faithfulModel)
## End(Not run)

[Package mclust version 5.4.2 Index]