diabetes {mclust}	R Documentation

Diabetes data

Description

The data set contains three measurements made on 145 non-obese adult patients classified into three groups.

Usage

data(diabetes)

Format

A data frame with the following variables:

class

The type of diabete: Normal, Overt, and Chemical.

glucose

Area under plasma glucose curve after a three hour oral glucose tolerance test (OGTT).

insulin

Area under plasma insulin curve after a three hour oral glucose tolerance test (OGTT).

sspg

Steady state plasma glucose.

Source

Reaven, G. M. and Miller, R. G. (1979). An attempt to define the nature of chemical diabetes using a multidimensional analysis. Diabetologia 16:17-24.

[Package mclust version 5.4.2 Index]