dupPartition {mclust}	R Documentation

Partition the data by grouping together duplicated data

Description

Duplicated data are grouped together to form a basic partition that can be used to start hierarchical agglomeration.

Usage

dupPartition(data)

Arguments

data
A numeric vector, matrix, or data frame of observations. If a matrix or data frame, rows correspond to observations (n) and columns correspond to variables (d).

Value

A vector of indices indicating the partition.

See Also

hc

Examples

dupPartition(iris[,1:4])
dupPartition(iris)
dupPartition(iris$Species)

[Package mclust version 5.4.9 Index]