estepE {mclust}	R Documentation

E-step in the EM algorithm for a parameterized Gaussian mixture model.

Description

Implements the expectation step in the EM algorithm for a parameterized Gaussian mixture model.

Usage

estepE(data, parameters, warn = NULL, ...)
estepV(data, parameters, warn = NULL, ...)
estepEII(data, parameters, warn = NULL, ...)
estepVII(data, parameters, warn = NULL, ...)
estepEEI(data, parameters, warn = NULL, ...)
estepVEI(data, parameters, warn = NULL, ...)
estepEVI(data, parameters, warn = NULL, ...)
estepVVI(data, parameters, warn = NULL, ...)
estepEEE(data, parameters, warn = NULL, ...)
estepEEV(data, parameters, warn = NULL, ...)
estepVEV(data, parameters, warn = NULL, ...)
estepVVV(data, parameters, warn = NULL, ...)
estepEVE(data, parameters, warn = NULL, ...)
estepEVV(data, parameters, warn = NULL, ...)
estepVEE(data, parameters, warn = NULL, ...)
estepVVE(data, parameters, warn = NULL, ...)

Arguments

data
A numeric vector, matrix, or data frame of observations. Categorical variables are not allowed. If a matrix or data frame, rows correspond to observations and columns correspond to variables.

parameters
The parameters of the model:

pro

Mixing proportions for the components of the mixture. If the model includes a Poisson term for noise, there should be one more mixing proportion than the number of Gaussian components.

mu

The mean for each component. If there is more than one component, this is a matrix whose columns are the means of the components.

variance

A list of variance parameters for the model. The components of this list depend on the model specification. See the help file for mclustVariance for details.

Vinv

An estimate of the reciprocal hypervolume of the data region. If not supplied or set to a negative value, the default is determined by applying function hypvol to the data. Used only when pro includes an additional mixing proportion for a noise component.

warn
A logical value indicating whether or certain warnings should be issued. The default is given by mclust.options("warn").

...
Catches unused arguments in indirect or list calls via do.call.

Value

A list including the following components:

modelName
Character string identifying the model.

z
A matrix whose [i,k]th entry is the conditional probability of the ith observation belonging to the kth component of the mixture.

parameters
The input parameters.

loglik
The logliklihood for the data in the mixture model.

Attribute
"WARNING": An appropriate warning if problems are encountered in the computations.

See Also

estep, em, mstep, do.call, mclustVariance, mclust.options.

Examples

## Not run: 
msEst <- mstepEII(data = iris[,-5], z = unmap(iris[,5]))
names(msEst)

estepEII(data = iris[,-5], parameters = msEst$parameters)
## End(Not run)

[Package mclust version 5.4.2 Index]