hcRandomPairs {mclust}	R Documentation

Random hierarchical structure

Description

Create a hierarchical structure using a random hierarchical partition of the data.

Usage

hcRandomPairs(data, seed = NULL, ...)

Arguments

data
A numeric matrix or data frame of observations. If a matrix or data frame, rows correspond to observations and columns correspond to variables.

seed
Optional single value, interpreted as an integer, specifying the seed for random partition.

...
Catches unused arguments in indirect or list calls via do.call.

Value

A numeric two-column matrix in which the ith row gives the minimum index for observations in each of the two clusters merged at the ith stage of a random agglomerative hierarchical clustering.

See Also

hc, hclass hcVVV

Examples

data <- iris[,1:4]
randPairs <- hcRandomPairs(data)
str(randPairs)
# start model-based clustering from a random partition
mod <- Mclust(data, initialization = list(hcPairs = randPairs))
summary(mod)

[Package mclust version 5.4.9 Index]