icl {mclust}	R Documentation

ICL for an estimated Gaussian Mixture Model

Description

Computes the ICL (Integrated Complete-data Likelihood) for criterion for a Gaussian Mixture Model fitted by Mclust.

Usage

icl(object, ...)

Arguments

object
An object of class 'Mclust' resulting from a call to Mclust.

...
Further arguments passed to or from other methods.

Value

The ICL for the given input MCLUST model.

References

Biernacki, C., Celeux, G., Govaert, G. (2000). Assessing a mixture model for clustering with the integrated completed likelihood. IEEE Trans. Pattern Analysis and Machine Intelligence, 22 (7), 719-725.

See Also

Mclust, mclustBIC, mclustICL, bic.

Examples

mod <- Mclust(iris[,1:4])
icl(mod)

[Package mclust version 5.4.2 Index]