mclust.options {mclust}	R Documentation

Default values for use with MCLUST package

Description

Set or retrieve default values for use with MCLUST package.

Usage

mclust.options(...)

Arguments

...
one or more arguments provided in the name = value form, or no argument at all may be given.
Available arguments are described in the Details section below.

Details

mclust.options is provided for assigning or retrieving default values used by various functions in MCLUST.

Available options are:

emModelNames

A vector of 3-character strings that are associated with multivariate models for which EM estimation is available in MCLUST.
The current default is all of the multivariate mixture models supported in MCLUST. The help file for mclustModelNames describes the available models.

hcModelName

A string associated with multivariate models for which model-based hierarchical clustering is available in MCLUST.
The available models are the following:

"EII" 	 spherical, equal volume

"EEE" 	 ellipsoidal, equal volume, shape, and orientation

"VII" 	 spherical, unequal volume

"VVV" 	 ellipsoidal, varying volume, shape, and orientation.

The "VVV" is used as default for initialization of EM algorithm.

hcUse

A string or a vector of character strings specifying the type of input variables to be used in model-based hierarchical clustering to start the EM algorithm. Possible values are:

"VARS" 	 original variables;

"STD" 	 standardized variables;

"SPH" 	 sphered variables (centered, scaled, uncorrelated) computed using SVD;

"PCS" 	 principal components computed using SVD on centered variables (i.e. using the covariance matrix);

"PCR" 	 principal components computed using SVD on standardized (center and scaled) variables (i.e. using the correlation matrix);

"SVD" 	 scaled SVD transformation;

"RND" 	 no transformation is applied but a random hierarchical structure is returned (see randomPairs).

For further details see Scrucca and Raftery (2015), Scrucca et al. (2016).

subset

A value specifying the maximal sample size to be used in the model-based hierarchical clustering to start the EM algorithm. If data sample size exceeds this value, a random sample is drawn of size specified by subset.

bicPlotSymbols

A vector whose entries correspond to graphics symbols for plotting the BIC values output from Mclust and mclustBIC. These are displayed in the legend which appears at the lower right of the BIC plots.

bicPlotColors

A vector whose entries correspond to colors for plotting the BIC curves from output from Mclust and mclustBIC. These are displayed in the legend which appears at the lower right of the BIC plots.

classPlotSymbols

A vector whose entries are either integers corresponding to graphics symbols or single characters for indicating classifications when plotting data. Classes are assigned symbols in the given order.

classPlotColors

A vector whose entries correspond to colors for indicating classifications when plotting data. Classes are assigned colors in the given order.

warn

A logical value indicating whether or not to issue certain warnings. Most of these warnings have to do with situations in which singularities are encountered. The default is warn = FALSE.

The parameter values set via a call to this function will remain in effect for the rest of the session, affecting the subsequent behaviour of the functions for which the given parameters are relevant.

Value

If the argument list is empty the function returns the current list of values. If the argument list is not empty, the returned list is invisible.

References

Scrucca L. and Raftery A. E. (2015) Improved initialisation of model-based clustering using Gaussian hierarchical partitions. Advances in Data Analysis and Classification, 9/4, pp. 447-460.

Scrucca L., Fop M., Murphy T. B. and Raftery A. E. (2016) mclust 5: clustering, classification and density estimation using Gaussian finite mixture models, The R Journal, 8/1, pp. 205-233.

See Also

Mclust, MclustDA, densityMclust, emControl

Examples

opt <- mclust.options() # save default values
irisBIC <- mclustBIC(iris[,-5])
summary(irisBIC, iris[,-5])

mclust.options(emModelNames = c("EII", "EEI", "EEE"))
irisBIC <- mclustBIC(iris[,-5])
summary(irisBIC, iris[,-5])

mclust.options(opt)    # restore default values
mclust.options()

oldpar <- par(mfrow = c(2,1), no.readonly = TRUE)
n <- with(mclust.options(), 
          max(sapply(list(bicPlotSymbols, bicPlotColors),length)))
plot(seq(n), rep(1,n), ylab = "", xlab = "", yaxt = "n", 
     pch = mclust.options("bicPlotSymbols"), 
     col = mclust.options("bicPlotColors"))
title("mclust.options(\"bicPlotSymbols\") \n mclust.options(\"bicPlotColors\")")
n <- with(mclust.options(), 
          max(sapply(list(classPlotSymbols, classPlotColors),length)))
plot(seq(n), rep(1,n), ylab = "", xlab = "", yaxt = "n", 
     pch = mclust.options("classPlotSymbols"), 
     col = mclust.options("classPlotColors"))
title("mclust.options(\"classPlotSymbols\") \n mclust.options(\"classPlotColors\")")
par(oldpar)

[Package mclust version 5.4.2 Index]