mclustBICupdate {mclust}	R Documentation

Update BIC values for parameterized Gaussian mixture models

Description

Update the BIC (Bayesian Information Criterion) for parameterized Gaussian mixture models by taking the best from BIC results as returned by mclustBIC.

Usage

mclustBICupdate(BIC, ...)

Arguments

BIC
Object of class 'mclustBIC' containing the BIC values as returned by a call to mclustBIC.

...
Further objects of class 'mclustBIC' to be merged.

Value

An object of class 'mclustBIC' containing the best values obtained from merging the input arguments. Attributes are also updated according to the best BIC found, so calling Mclust on the resulting ouput will return the corresponding best model (see example).

See Also

mclustBIC, Mclust.

Examples

## Not run: 
data(galaxies, package = "MASS") 
galaxies <- galaxies / 1000

# use several random starting points
BIC <- NULL
for(j in 1:100)
{
  rBIC <- mclustBIC(galaxies, verbose = FALSE,
                    initialization = list(hcPairs = randomPairs(galaxies)))
  BIC <- mclustBICupdate(BIC, rBIC)
}
pickBIC(BIC)
plot(BIC)

mod <- Mclust(galaxies, x = BIC)
summary(mod)

## End(Not run)

[Package mclust version 5.4.2 Index]