mclustLoglik {mclust}	R Documentation

Log-likelihood from a table of BIC values for parameterized Gaussian mixture models

Description

Compute the maximal log-likelihood from a table of BIC values contained in a 'mclustBIC' object as returned by function mclustBIC.

Usage

mclustLoglik(object, ...)

Arguments

object
An object of class 'mclustBIC' containing the BIC values as returned by a call to mclustBIC.

...
Catches unused arguments in an indirect or list call via do.call.

Value

An object of class 'mclustLoglik' containing the maximal log-likelihood values for the Gaussian mixture models provided as input.

See Also

mclustBIC.

Examples

## Not run: 
BIC <- mclustBIC(iris[,1:4])
mclustLoglik(BIC)

## End(Not run)

[Package mclust version 5.4.2 Index]