me {mclust}	R Documentation

EM algorithm starting with M-step for parameterized MVN mixture models.

Description

Implements the EM algorithm for MVN mixture models parameterized by eignevalue decomposition, starting with the maximization step.

Usage

me(modelName, data, z, prior = NULL, control = emControl(), 
   Vinv = NULL, warn = NULL, ...)

Arguments

modelName
A character string indicating the model. The help file for mclustModelNames describes the available models.

data
A numeric vector, matrix, or data frame of observations. Categorical variables are not allowed. If a matrix or data frame, rows correspond to observations and columns correspond to variables.

z
A matrix whose [i,k]th entry is an initial estimate of the conditional probability of the ith observation belonging to the kth component of the mixture.

prior
Specification of a conjugate prior on the means and variances. See the help file for priorControl for further information. The default assumes no prior.

control
A list of control parameters for EM. The defaults are set by the call emControl().

Vinv
If the model is to include a noise term, Vinv is an estimate of the reciprocal hypervolume of the data region. If set to a negative value or 0, the model will include a noise term with the reciprocal hypervolume estimated by the function hypvol. The default is not to assume a noise term in the model through the setting Vinv=NULL.

warn
A logical value indicating whether or not certain warnings (usually related to singularity) should be issued when the estimation fails. The default is set in mclust.options("warn").

...
Catches unused arguments in indirect or list calls via do.call.

Value

A list including the following components:

modelName
A character string identifying the model (same as the input argument).

n
The number of observations in the data.

d
The dimension of the data.

G
The number of mixture components.

z
A matrix whose [i,k]th entry is the conditional probability of the ith observation belonging to the kth component of the mixture.

parameters

pro

A vector whose kth component is the mixing proportion for the kth component of the mixture model. If the model includes a Poisson term for noise, there should be one more mixing proportion than the number of Gaussian components.

mean

The mean for each component. If there is more than one component, this is a matrix whose kth column is the mean of the kth component of the mixture model.

variance

A list of variance parameters for the model. The components of this list depend on the model specification. See the help file for mclustVariance for details.

Vinv

The estimate of the reciprocal hypervolume of the data region used in the computation when the input indicates the addition of a noise component to the model.

loglik
The log likelihood for the data in the mixture model.

control
The list of control parameters for EM used.

prior
The specification of a conjugate prior on the means and variances used, NULL if no prior is used.

Attributes:
"info" Information on the iteration.
"WARNING" An appropriate warning if problems are encountered in the computations.

See Also

meE,..., meVVV, em, mstep, estep, priorControl, mclustModelNames, mclustVariance, mclust.options

Examples

## Not run: 
me(modelName = "VVV", data = iris[,-5], z = unmap(iris[,5]))
## End(Not run)

[Package mclust version 5.4.2 Index]