mstep {mclust}	R Documentation

M-step for parameterized Gaussian mixture models.

Description

Maximization step in the EM algorithm for parameterized Gaussian mixture models.

Usage

mstep(modelName, data, z, prior = NULL, warn = NULL, ...)

Arguments

modelName
A character string indicating the model. The help file for mclustModelNames describes the available models.

data
A numeric vector, matrix, or data frame of observations. Categorical variables are not allowed. If a matrix or data frame, rows correspond to observations and columns correspond to variables.

z
A matrix whose [i,k]th entry is the conditional probability of the ith observation belonging to the kth component of the mixture. In analyses involving noise, this should not include the conditional probabilities for the noise component.

prior
Specification of a conjugate prior on the means and variances. The default assumes no prior.

warn
A logical value indicating whether or not certain warnings (usually related to singularity) should be issued when the estimation fails. The default is given by mclust.options("warn").

...
Catches unused arguments in indirect or list calls via do.call.

Value

A list including the following components:

modelName
A character string identifying the model (same as the input argument).

parameters

pro

A vector whose kth component is the mixing proportion for the kth component of the mixture model. If the model includes a Poisson term for noise, there should be one more mixing proportion than the number of Gaussian components.

mean

The mean for each component. If there is more than one component, this is a matrix whose kth column is the mean of the kth component of the mixture model.

variance

A list of variance parameters for the model. The components of this list depend on the model specification. See the help file for mclustVariance for details.

Attributes:
"info" For those models with iterative M-steps ("VEI" and "VEV"), information on the iteration.
"WARNING" An appropriate warning if problems are encountered in the computations.

Note

This function computes the M-step only for MVN mixtures, so in analyses involving noise, the conditional probabilities input should exclude those for the noise component.

In contrast to me for the EM algorithm, computations in mstep are carried out unless failure due to overflow would occur. To impose stricter tolerances on a single mstep, use me with the itmax component of the control argument set to 1.

See Also

mstepE, ..., mstepVVV, emControl, me, estep, mclust.options.

Examples

## Not run: 
mstep(modelName = "VII", data = iris[,-5], z = unmap(iris[,5]))
## End(Not run)

[Package mclust version 5.4.2 Index]