nVarParams {mclust}	R Documentation

Number of Variance Parameters in Gaussian Mixture Models

Description

Gives the number of variance parameters for parameterizations of the Gaussian mixture model that are used in MCLUST.

Usage

nVarParams(modelName, d, G, ...)

Arguments

modelName
A character string indicating the model. The help file for mclustModelNames describes the available models.

d
The dimension of the data. Not used for models in which neither the shape nor the orientation varies.

G
The number of components in the Gaussian mixture model used to compute loglik.

...
Catches unused arguments in indirect or list calls via do.call.

Details

To get the total number of parameters in model, add G*d for the means and G-1 for the mixing proportions if they are unequal.

Value

The number of variance parameters in the corresponding Gaussian mixture model.

References

C. Fraley and A. E. Raftery (2002). Model-based clustering, discriminant analysis, and density estimation. Journal of the American Statistical Association 97:611:631.

C. Fraley, A. E. Raftery, T. B. Murphy and L. Scrucca (2012). mclust Version 4 for R: Normal Mixture Modeling for Model-Based Clustering, Classification, and Density Estimation. Technical Report No. 597, Department of Statistics, University of Washington.

See Also

bic, nMclustParams.

Examples

mapply(nVarParams, mclust.options("emModelNames"), d = 2, G = 3)

[Package mclust version 5.4.2 Index]