plot.MclustSSC {mclust}	R Documentation

Plotting method for MclustSSC semi-supervised classification

Description

Plots for semi-supervised classification based on Gaussian finite mixture models.

Usage

## S3 method for class 'MclustSSC'
plot(x, what = c("BIC", "classification", "uncertainty"), ...)

Arguments

x
An object of class 'MclustSSC' resulting from a call to MclustSSC.

what
A string specifying the type of graph requested. Available choices are:

"BIC" =

plot of BIC values used for model selection, i.e. for choosing the model class covariances.

"classification" =

a plot of data with points marked based on the known and the predicted classification.

"uncertainty" =

a plot of classification uncertainty.

If not specified, in interactive sessions a menu of choices is proposed.

...
further arguments passed to or from other methods. See plot.Mclust.

Author(s)

Luca Scrucca

See Also

MclustSSC

Examples

X <- iris[,1:4]
class <- iris$Species
# randomly remove class labels
set.seed(123)
class[sample(1:length(class), size = 120)] <- NA
table(class, useNA = "ifany")
clPairs(X, ifelse(is.na(class), 0, class),
        symbols = c(0, 16, 17, 18), colors = c("grey", 4, 2, 3),
        main = "Partially classified data")

# Fit semi-supervised classification model
mod_SSC  <- MclustSSC(X, class)
summary(mod_SSC, parameters = TRUE)

pred_SSC <- predict(mod_SSC)
table(Predicted = pred_SSC$classification, Actual = class, useNA = "ifany")

plot(mod_SSC, what = "BIC")
plot(mod_SSC, what = "classification")
plot(mod_SSC, what = "uncertainty")

[Package mclust version 5.4.9 Index]