predict.Mclust {mclust}	R Documentation

Cluster multivariate observations by Gaussian finite mixture modeling

Description

Cluster prediction for multivariate observations based on Gaussian finite mixture models estimated by Mclust.

Usage

  ## S3 method for class 'Mclust'
predict(object, newdata, ...)

Arguments

object
an object of class 'Mclust' resulting from a call to Mclust.

newdata
a data frame or matrix giving the data. If missing the clustering data obtained from the call to Mclust are classified.

...
further arguments passed to or from other methods.

Value

Returns a list of with the following components:

classification
a factor of predicted cluster labels for newdata.

z
a matrix whose [i,k]th entry is the probability that observation i in newdata belongs to the kth cluster.

Author(s)

Luca Scrucca

See Also

Mclust.

Examples

model <- Mclust(faithful)

# predict cluster for the observed data
pred <- predict(model) 
str(pred)
pred$z              # equal to model$z
pred$classification # equal to 
plot(faithful, col = pred$classification, pch = pred$classification)

# predict cluster over a grid
grid <- apply(faithful, 2, function(x) seq(min(x), max(x), length = 50))
grid <- expand.grid(eruptions = grid[,1], waiting = grid[,2])
pred <- predict(model, grid)
plot(grid, col = mclust.options("classPlotColors")[pred$classification], pch = 15, cex = 0.5)
points(faithful, pch = model$classification)

[Package mclust version 5.4.2 Index]