predict.MclustDA {mclust}	R Documentation

Classify multivariate observations by Gaussian finite mixture modeling

Description

Classify multivariate observations based on Gaussian finite mixture models estimated by MclustDA.

Usage

  ## S3 method for class 'MclustDA'
predict(object, newdata, prior, ...)

Arguments

object
an object of class 'MclustDA' resulting from a call to MclustDA.

newdata
a data frame or matrix giving the data. If missing the train data obtained from the call to MclustDA are classified.

prior
the prior probabilities of the classes; by default, this is set at the proportions in the training data.

...
further arguments passed to or from other methods.

Value

Returns a list of with the following components:

classification
a factor of predicted class labels for newdata.

z
a matrix whose [i,k]th entry is the probability that observation i in newdata belongs to the kth class.

Author(s)

Luca Scrucca

See Also

MclustDA.

Examples

## Not run: 
odd <- seq(from = 1, to = nrow(iris), by = 2)
even <- odd + 1
X.train <- iris[odd,-5]
Class.train <- iris[odd,5]
X.test <- iris[even,-5]
Class.test <- iris[even,5]

irisMclustDA <- MclustDA(X.train, Class.train)

predTrain <- predict(irisMclustDA)
predTrain
predTest <- predict(irisMclustDA, X.test)
predTest

## End(Not run)

[Package mclust version 5.4.2 Index]