priorControl {mclust}	R Documentation

Conjugate Prior for Gaussian Mixtures.

Description

Specify a conjugate prior for Gaussian mixtures.

Usage

priorControl(functionName = "defaultPrior", ...) 

Arguments

functionName
The name of the function specifying the conjugate prior. By default the function defaultPrior is used, and this can also be used as a template for alternative specification.

...
Optional named arguments to the function specified in functionName together with their values.

Details

The function priorControl is used to specify a conjugate prior for EM within MCLUST.
Note that, as described in defaultPrior, in the multivariate case only 10 out of 14 models may be used in conjunction with a prior, i.e. those available in MCLUST up to version 4.4.

Value

A list with the function name as the first component. The remaining components (if any) consist of a list of arguments to the function with assigned values.

References

C. Fraley and A. E. Raftery (2007). Bayesian regularization for normal mixture estimation and model-based clustering. Journal of Classification 24:155-181.

See Also

mclustBIC, me, mstep, defaultPrior

Examples

# default prior
irisBIC <- mclustBIC(iris[,-5], prior = priorControl())
summary(irisBIC, iris[,-5])

# no prior on the mean; default prior on variance
irisBIC <- mclustBIC(iris[,-5], prior = priorControl(shrinkage = 0))
summary(irisBIC, iris[,-5])

[Package mclust version 5.4.2 Index]