Summarizing Gaussian Finite Mixture Model Fits

Summary method for class "Mclust".

# S3 method for Mclust
summary(object, parameters = FALSE, classification = FALSE, …)
# S3 method for summary.Mclust
print(x, digits = getOption("digits"), …)

Arguments

object
An object of class 'Mclust' resulting of a call to Mclust or densityMclust.

x
An object of class 'summary.Mclust', usually, a result of a call to summary.Mclust.

parameters
Logical; if TRUE, the parameters of mixture components are printed.

classification
Logical; if TRUE, the MAP classification/clustering of observations is printed.

digits
The number of significant digits to use when printing.

…
Further arguments passed to or from other methods.

See also

Examples

mod1 = Mclust(iris[,1:4])
summary(mod1)

#> ---------------------------------------------------- 
#> Gaussian finite mixture model fitted by EM algorithm 
#> ---------------------------------------------------- 
#> 
#> Mclust VEV (ellipsoidal, equal shape) model with 2 components: 
#> 
#>  log-likelihood   n df       BIC       ICL
#>        -215.726 150 26 -561.7285 -561.7289
#> 
#> Clustering table:
#>   1   2 
#>  50 100 

summary(mod1, parameters = TRUE, classification = TRUE)

#> ---------------------------------------------------- 
#> Gaussian finite mixture model fitted by EM algorithm 
#> ---------------------------------------------------- 
#> 
#> Mclust VEV (ellipsoidal, equal shape) model with 2 components: 
#> 
#>  log-likelihood   n df       BIC       ICL
#>        -215.726 150 26 -561.7285 -561.7289
#> 
#> Clustering table:
#>   1   2 
#>  50 100 
#> 
#> Mixing probabilities:
#>         1         2 
#> 0.3333319 0.6666681 
#> 
#> Means:
#>                   [,1]     [,2]
#> Sepal.Length 5.0060022 6.261996
#> Sepal.Width  3.4280049 2.871999
#> Petal.Length 1.4620007 4.905992
#> Petal.Width  0.2459998 1.675997
#> 
#> Variances:
#> [,,1]
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length   0.15065114  0.13080115   0.02084463  0.01309107
#> Sepal.Width    0.13080115  0.17604529   0.01603245  0.01221458
#> Petal.Length   0.02084463  0.01603245   0.02808260  0.00601568
#> Petal.Width    0.01309107  0.01221458   0.00601568  0.01042365
#> [,,2]
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length    0.4000438  0.10865444    0.3994018  0.14368256
#> Sepal.Width     0.1086544  0.10928077    0.1238904  0.07284384
#> Petal.Length    0.3994018  0.12389040    0.6109024  0.25738990
#> Petal.Width     0.1436826  0.07284384    0.2573899  0.16808182

mod2 = Mclust(iris[,1:4], G = 1)
summary(mod2, parameters = TRUE, classification = TRUE)

#> ---------------------------------------------------- 
#> Gaussian finite mixture model fitted by EM algorithm 
#> ---------------------------------------------------- 
#> 
#> Mclust XXX (ellipsoidal multivariate normal) model with 1 component: 
#> 
#>  log-likelihood   n df       BIC       ICL
#>       -379.9146 150 14 -829.9782 -829.9782
#> 
#> Clustering table:
#>   1 
#> 150 
#> 
#> Mixing probabilities:
#> 1 
#> 1 
#> 
#> Means:
#>                  [,1]
#> Sepal.Length 5.843333
#> Sepal.Width  3.057333
#> Petal.Length 3.758000
#> Petal.Width  1.199333
#> 
#> Variances:
#> [,,1]
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length   0.68112222 -0.04215111    1.2658200   0.5128289
#> Sepal.Width   -0.04215111  0.18871289   -0.3274587  -0.1208284
#> Petal.Length   1.26582000 -0.32745867    3.0955027   1.2869720
#> Petal.Width    0.51282889 -0.12082844    1.2869720   0.5771329

mod3 = Mclust(iris[,1:4], prior = priorControl())

#> Warning: The presence of BIC values equal to NA is likely due to one or more of the mixture proportions being estimated as zero, so that the model estimated reduces to one with a smaller number of components.

summary(mod3)

#> ---------------------------------------------------- 
#> Gaussian finite mixture model fitted by EM algorithm 
#> ---------------------------------------------------- 
#> 
#> Mclust VEV (ellipsoidal, equal shape) model with 2 components: 
#> 
#> Prior: defaultPrior() 
#> 
#>  log-likelihood   n df       BIC      ICL
#>       -225.2685 150 26 -580.8136 -580.814
#> 
#> Clustering table:
#>   1   2 
#>  50 100 

mod4 = Mclust(iris[,1:4], prior = priorControl(functionName="defaultPrior", shrinkage=0.1))

#> Warning: The presence of BIC values equal to NA is likely due to one or more of the mixture proportions being estimated as zero, so that the model estimated reduces to one with a smaller number of components.

summary(mod4, parameters = TRUE, classification = TRUE)

#> ---------------------------------------------------- 
#> Gaussian finite mixture model fitted by EM algorithm 
#> ---------------------------------------------------- 
#> 
#> Mclust VEV (ellipsoidal, equal shape) model with 2 components: 
#> 
#> Prior: defaultPrior(shrinkage = 0.1) 
#> 
#>  log-likelihood   n df       BIC       ICL
#>       -227.6729 150 26 -585.6223 -585.6227
#> 
#> Clustering table:
#>   1   2 
#>  50 100 
#> 
#> Mixing probabilities:
#>         1         2 
#> 0.3333319 0.6666681 
#> 
#> Means:
#>                   [,1]     [,2]
#> Sepal.Length 5.0076734 6.261578
#> Sepal.Width  3.4272648 2.872184
#> Petal.Length 1.4665835 4.904845
#> Petal.Width  0.2479027 1.675521
#> 
#> Variances:
#> [,,1]
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length   0.20575793  0.17025370   0.05542370  0.02870217
#> Sepal.Width    0.17025370  0.21011277   0.03926788  0.02243701
#> Petal.Length   0.05542370  0.03926788   0.06789947  0.02333949
#> Petal.Width    0.02870217  0.02243701   0.02333949  0.02037413
#> [,,2]
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length    0.4048913  0.10410868    0.3671040  0.11699659
#> Sepal.Width     0.1041087  0.10356921    0.1117050  0.06496733
#> Petal.Length    0.3671040  0.11170501    0.5942949  0.25703252
#> Petal.Width     0.1169966  0.06496733    0.2570325  0.17424925