summary.MclustDA {mclust}	R Documentation

Summarizing discriminant analysis based on Gaussian finite mixture modeling.

Description

Summary method for class "MclustDA".

Usage

## S3 method for class 'MclustDA'
summary(object, parameters = FALSE, newdata, newclass, ...)
## S3 method for class 'summary.MclustDA'
print(x, digits = getOption("digits"), ...)

Arguments

object
An object of class 'MclustDA' resulting from a call to MclustDA.

x
An object of class 'summary.MclustDA', usually, a result of a call to summary.MclustDA.

parameters
Logical; if TRUE, the parameters of mixture components are printed.

newdata
A data frame or matrix giving the test data.

newclass
A vector giving the class labels for the observations in the test data.

digits
The number of significant digits to use when printing.

...
Further arguments passed to or from other methods.

Value

The function summary.MclustDA computes and returns a list of summary statistics of the estimated MclustDA or EDDA model for classification.

Author(s)

Luca Scrucca

See Also

MclustDA, plot.MclustDA.

Examples

mod = MclustDA(data = iris[,1:4], class = iris$Species)
summary(mod)
summary(mod, parameters = TRUE)

[Package mclust version 5.4.2 Index]