summary.MclustDR {mclust}	R Documentation

Summarizing dimension reduction method for model-based clustering and classification

Description

Summary method for class "MclustDR".

Usage

## S3 method for class 'MclustDR'
summary(object, numdir, std = FALSE, ...)
## S3 method for class 'summary.MclustDR'
print(x, digits = max(5, getOption("digits") - 3), ...)

Arguments

object
An object of class 'MclustDR' resulting from a call to MclustDR.

x
An object of class 'summary.MclustDR', usually, a result of a call to summary.MclustDR.

numdir
An integer providing the number of basis directions to be printed.

std
if TRUE the coefficients basis are scaled such that all predictors have unit standard deviation.

digits
The number of significant digits to use when printing.

...
Further arguments passed to or from other methods.

Author(s)

Luca Scrucca

See Also

MclustDR, plot.MclustDR

[Package mclust version 5.4.2 Index]