Component Density for a Parameterized MVN Mixture Model

Computes component densities for points in a parameterized MVN mixture model.

cdensE(data, logarithm = FALSE, parameters, warn = NULL, ...)
cdensV(data, logarithm = FALSE, parameters, warn = NULL, ...)
cdensX(data, logarithm = FALSE, parameters, warn = NULL, ...)
cdensEII(data, logarithm = FALSE, parameters, warn = NULL, ...)
cdensVII(data, logarithm = FALSE, parameters, warn = NULL, ...)
cdensEEI(data, logarithm = FALSE, parameters, warn = NULL, ...)
cdensVEI(data, logarithm = FALSE, parameters, warn = NULL, ...)
cdensEVI(data, logarithm = FALSE, parameters, warn = NULL, ...)
cdensVVI(data, logarithm = FALSE, parameters, warn = NULL, ...)
cdensEEE(data, logarithm = FALSE, parameters, warn = NULL, ...)
cdensEEV(data, logarithm = FALSE, parameters, warn = NULL, ...)
cdensVEV(data, logarithm = FALSE, parameters, warn = NULL, ...)
cdensVVV(data, logarithm = FALSE, parameters, warn = NULL, ...)
cdensEVE(data, logarithm = FALSE, parameters, warn = NULL, ...)
cdensEVV(data, logarithm = FALSE, parameters, warn = NULL, ...)
cdensVEE(data, logarithm = FALSE, parameters, warn = NULL, ...)
cdensVVE(data, logarithm = FALSE, parameters, warn = NULL, ...) 
cdensXII(data, logarithm = FALSE, parameters, warn = NULL, ...)
cdensXXI(data, logarithm = FALSE, parameters, warn = NULL, ...)
cdensXXX(data, logarithm = FALSE, parameters, warn = NULL, ...)

Arguments

data

A numeric vector, matrix, or data frame of observations. Categorical variables are not allowed. If a matrix or data frame, rows correspond to observations and columns correspond to variables.

logarithm

A logical value indicating whether or not the logarithm of the component densities should be returned. The default is to return the component densities, obtained from the log component densities by exponentiation.

parameters

The parameters of the model:

mean

The mean for each component. If there is more than one component, this is a matrix whose kth column is the mean of the kth component of the mixture model.

variance

A list of variance parameters for the model. The components of this list depend on the model specification. See the help file for mclustVariance for details.

pro

Mixing proportions for the components of the mixture. If the model includes a Poisson term for noise, there should be one more mixing proportion than the number of Gaussian components.

warn

A logical value indicating whether or not a warning should be issued when computations fail. The default is warn=FALSE.

...

Catches unused arguments in indirect or list calls via do.call.

Value

A numeric matrix whose [i,j]th entry is the density of observation i in component j. The densities are not scaled by mixing proportions.

Note

When one or more component densities are very large in magnitude, then it may be possible to compute the logarithm of the component densities but not the component densities themselves due to overflow.

See also

Examples

# \donttest{
z2 <- unmap(hclass(hcVVV(faithful),2)) # initial value for 2 class case

model <- meVVV(data=faithful, z=z2)
cdensVVV(data=faithful, logarithm = TRUE, parameters = model$parameters)
#>           [,1]       [,2]
#> 1    -4.197962 -23.327906
#> 2   -23.342333  -2.639804
#> 3    -5.369492 -16.425216
#> 4   -15.294924  -3.232768
#> 5    -3.060038 -50.710265
#> 6   -13.290054  -7.761979
#> 7    -3.699704 -57.918268
#> 8    -5.334659 -26.752622
#> 9   -22.962876  -2.412796
#> 10   -3.034001 -44.933104
#> 11  -22.934312  -2.528941
#> 12   -3.676302 -32.771717
#> 13   -2.724501 -37.562422
#> 14  -27.479203  -3.332165
#> 15   -3.170498 -55.618395
#> 16  -20.214896  -2.518439
#> 17  -21.915710  -4.216529
#> 18   -3.455892 -59.675676
#> 19  -26.760842  -3.595670
#> 20   -2.679311 -39.367492
#> 21  -24.676501  -2.687279
#> 22  -27.479203  -3.332165
#> 23   -4.858440 -20.080789
#> 24   -7.357413 -10.919858
#> 25   -3.707075 -47.408698
#> 26   -4.826034 -25.481836
#> 27  -20.944700  -2.269342
#> 28   -2.921123 -33.631614
#> 29   -3.238694 -28.304652
#> 30   -2.776193 -45.080634
#> 31   -3.466375 -39.469218
#> 32   -3.010859 -45.669805
#> 33   -6.449804 -15.147156
#> 34   -2.893828 -33.676595
#> 35   -3.472704 -26.585813
#> 36  -21.731435  -2.312102
#> 37  -25.554020  -2.900109
#> 38   -3.677658 -59.639204
#> 39  -21.329762  -3.062301
#> 40   -4.247178 -61.991292
#> 41   -2.677669 -42.761619
#> 42  -20.946033  -2.714943
#> 43   -2.993015 -51.386831
#> 44  -22.685815  -3.270025
#> 45   -3.960485 -47.302375
#> 46   -6.598464 -20.802812
#> 47   -6.211837 -25.551593
#> 48  -20.374670  -2.305713
#> 49   -3.012673 -52.862464
#> 50  -19.230203  -2.593682
#> 51   -4.415900 -57.463454
#> 52   -4.155687 -59.620152
#> 53  -22.934312  -2.528941
#> 54   -3.677658 -59.639204
#> 55  -24.193790  -2.917515
#> 56   -3.702828 -62.476248
#> 57   -4.178361 -23.158393
#> 58  -22.961895  -5.472180
#> 59   -3.220395 -49.108148
#> 60   -2.680246 -42.118651
#> 61  -16.621517  -2.671928
#> 62   -2.929064 -49.161225
#> 63  -26.881283  -3.164999
#> 64   -3.442207 -58.957948
#> 65  -21.320242  -3.337036
#> 66   -4.816504 -50.784915
#> 67   -2.736993 -36.607377
#> 68   -3.452079 -54.126106
#> 69  -17.334988  -3.949120
#> 70   -4.544447 -53.441400
#> 71   -3.052678 -34.577311
#> 72  -20.567346  -2.316439
#> 73   -2.869289 -47.303898
#> 74   -3.793379 -30.306906
#> 75  -18.780967  -3.237428
#> 76   -5.551980 -68.561758
#> 77  -18.770649  -2.736682
#> 78   -3.090617 -49.336764
#> 79   -3.201066 -28.392013
#> 80   -4.826034 -25.481836
#> 81   -3.011010 -34.787362
#> 82   -2.723047 -43.012954
#> 83   -4.062219 -33.036185
#> 84  -11.298804  -5.524240
#> 85   -3.345808 -32.488922
#> 86   -4.208793 -66.504258
#> 87   -3.076628 -30.077285
#> 88   -2.841868 -48.182626
#> 89  -22.581164  -3.203542
#> 90   -3.857276 -35.951643
#> 91  -16.678900  -2.737894
#> 92   -4.225630 -47.419501
#> 93  -24.413541  -2.615064
#> 94   -3.870686 -58.569039
#> 95  -20.629889  -4.070450
#> 96   -3.710463 -39.347779
#> 97   -3.145918 -54.839806
#> 98   -3.587831 -24.925274
#> 99  -23.891950  -2.520996
#> 100  -3.791269 -62.817662
#> 101 -13.314929  -4.015144
#> 102  -3.620549 -47.127765
#> 103 -22.563187  -2.807059
#> 104  -2.852246 -48.725153
#> 105  -2.924853 -34.554491
#> 106 -26.172906  -3.091087
#> 107  -3.211476 -56.013763
#> 108 -24.406371  -2.696018
#> 109  -3.729394 -62.392134
#> 110  -4.061563 -25.984956
#> 111  -4.136478 -54.869882
#> 112 -15.940531  -2.852310
#> 113  -4.281280 -65.752555
#> 114  -2.758530 -44.560136
#> 115 -23.139134  -3.748921
#> 116  -3.024552 -52.518067
#> 117 -20.025202  -3.417245
#> 118  -3.127196 -52.956969
#> 119 -21.540990  -3.130198
#> 120  -3.362336 -48.054637
#> 121 -16.003038  -5.076151
#> 122  -4.349739 -32.007303
#> 123  -2.792823 -38.748925
#> 124 -20.567346  -2.316439
#> 125  -3.595774 -54.494859
#> 126  -3.719606 -27.759651
#> 127 -26.996373  -3.552868
#> 128  -2.807859 -48.321385
#> 129 -17.801948  -2.616670
#> 130  -4.095762 -57.353351
#> 131 -27.507973  -3.569953
#> 132  -2.932976 -38.642050
#> 133 -13.219165  -6.598774
#> 134  -3.924297 -46.755621
#> 135 -27.188030  -3.355339
#> 136  -2.728314 -44.556354
#> 137 -23.709163  -2.491718
#> 138  -3.983323 -65.582239
#> 139 -21.075819  -2.254025
#> 140  -3.650663 -26.069747
#> 141  -2.704316 -39.618171
#> 142 -16.326223  -2.794102
#> 143  -2.847325 -49.422198
#> 144  -4.045263 -58.391303
#> 145  -2.958232 -41.064783
#> 146 -19.435158  -2.621356
#> 147  -3.068862 -52.205973
#> 148 -23.376850  -2.685890
#> 149  -6.763224 -77.891170
#> 150 -23.754624  -2.623325
#> 151  -5.104238 -67.202584
#> 152  -2.943287 -31.692611
#> 153 -13.516733  -4.266377
#> 154  -2.956467 -51.358696
#> 155  -4.605864 -19.879599
#> 156  -4.048571 -30.189017
#> 157  -2.795904 -47.949919
#> 158  -6.048704 -43.036756
#> 159 -23.754624  -2.623325
#> 160  -4.868659 -37.175298
#> 161 -24.424929  -4.196397
#> 162  -3.473474 -39.792856
#> 163 -19.516176  -2.456391
#> 164  -3.285133 -27.903995
#> 165  -5.999115 -17.754059
#> 166  -3.425381 -49.479805
#> 167 -14.207790  -3.676640
#> 168  -4.424334 -69.131072
#> 169 -22.648320  -2.350947
#> 170  -5.022950 -58.257673
#> 171 -24.420125  -2.683646
#> 172 -18.892222  -2.315269
#> 173  -3.260305 -49.672861
#> 174  -6.086198 -14.806481
#> 175  -2.757288 -37.731271
#> 176  -2.681165 -42.607500
#> 177  -3.867798 -46.141116
#> 178 -19.239233  -4.027522
#> 179  -3.625995 -35.349204
#> 180  -3.163973 -35.561099
#> 181 -21.935642  -2.425455
#> 182  -3.260305 -49.672861
#> 183  -2.841467 -40.992268
#> 184  -4.038653 -28.833247
#> 185 -22.081202  -2.412819
#> 186  -2.849526 -44.792442
#> 187  -3.225392 -36.894643
#> 188 -27.188030  -3.355339
#> 189  -2.800981 -46.048960
#> 190 -18.619497  -2.376900
#> 191  -3.484011 -58.647539
#> 192 -21.874287  -2.752047
#> 193  -4.179506 -57.580043
#> 194  -3.189917 -37.341307
#> 195  -2.996970 -30.808798
#> 196  -2.704316 -39.618171
#> 197  -6.608032 -26.436920
#> 198  -2.869070 -42.356538
#> 199 -20.000070  -2.924230
#> 200  -3.351040 -52.911852
#> 201 -17.793378  -2.672037
#> 202  -2.722907 -43.533303
#> 203  -5.032900 -42.701164
#> 204 -22.946061  -2.429771
#> 205  -3.168950 -50.499176
#> 206 -27.737489  -3.448581
#> 207  -2.870129 -42.388558
#> 208  -3.912003 -31.230266
#> 209 -24.248577  -2.673448
#> 210  -2.852246 -48.725153
#> 211 -13.414135  -6.405328
#> 212  -3.242152 -54.626111
#> 213 -24.967565  -2.741435
#> 214  -3.377164 -26.866904
#> 215  -6.924280 -15.976850
#> 216  -2.891784 -37.973694
#> 217 -17.613974  -3.401969
#> 218  -5.470209 -65.202518
#> 219 -20.568685  -2.238339
#> 220  -2.888887 -35.527396
#> 221 -24.413541  -2.615064
#> 222  -2.742433 -41.035802
#> 223 -23.974824  -2.840374
#> 224  -3.366504 -45.814869
#> 225  -2.915115 -32.036623
#> 226  -2.753490 -35.545441
#> 227  -2.802601 -34.253551
#> 228  -2.722189 -39.554079
#> 229  -4.089498 -27.945174
#> 230  -2.958889 -49.009002
#> 231  -4.055045 -32.541083
#> 232 -16.953218  -3.397325
#> 233  -3.409814 -40.685358
#> 234 -20.879977  -2.964063
#> 235  -4.097194 -50.902135
#> 236 -22.330375  -2.393566
#> 237 -22.727043  -2.478507
#> 238  -2.805091 -39.753714
#> 239  -3.018993 -31.139851
#> 240 -14.374714  -3.799218
#> 241  -3.008801 -35.278518
#> 242 -21.777249  -4.372194
#> 243  -3.983323 -65.582239
#> 244  -9.754270  -7.750691
#> 245  -3.107232 -52.380234
#> 246  -3.616466 -29.739030
#> 247 -18.892222  -2.315269
#> 248  -2.724757 -44.058193
#> 249 -16.384450  -4.578577
#> 250  -3.320311 -41.221848
#> 251 -18.901661  -2.449105
#> 252  -2.815687 -47.100028
#> 253  -4.347960 -20.388537
#> 254  -3.867798 -46.141116
#> 255  -3.979570 -41.033588
#> 256  -3.437195 -28.390477
#> 257  -3.849179 -28.079956
#> 258  -2.815687 -47.100028
#> 259 -20.185417  -2.278720
#> 260  -2.679752 -40.358850
#> 261  -3.680269 -56.644036
#> 262  -2.956702 -50.248606
#> 263 -21.366355  -2.826749
#> 264  -2.841467 -40.992268
#> 265 -27.853767  -4.245748
#> 266 -16.147464  -2.829734
#> 267  -4.204452 -55.521276
#> 268  -2.817372 -36.347359
#> 269 -24.104705  -3.678010
#> 270  -4.123883 -49.898080
#> 271 -27.361985  -3.380903
#> 272  -3.541020 -45.116720
#> attr(,"logarithm")
#> [1] TRUE
#> attr(,"modelName")
#> [1] "VVV"
#> attr(,"returnCode")
#> [1] 0

data(cross)
z2 <- unmap(cross[,1])

model <- meEEV(data = cross[,-1], z = z2)

EEVdensities <- cdensEEV( data = cross[,-1], parameters = model$parameters)

cbind(cross[,-1],map(EEVdensities))# }
#>                X1            X2 map(EEVdensities)
#> 1     1.262954285 -14.370462129                 1
#> 2    -0.326233361   4.418706353                 1
#> 3     1.329799263   3.794430288                 1
#> 4     1.272429321  16.865135087                 1
#> 5     0.414641434   9.310628915                 1
#> 6    -1.539950042   0.736292793                 2
#> 7    -0.928567035  -0.742713858                 2
#> 8    -0.294720447   5.454660878                 1
#> 9    -0.005767173  -7.986781308                 1
#> 10    2.404653389   0.948792512                 2
#> 11    0.763593461   3.175870260                 1
#> 12   -0.799009249   4.953540226                 1
#> 13   -1.147657009 -10.208978717                 1
#> 14   -0.289461574  13.161163849                 1
#> 15   -0.299215118   6.319050396                 1
#> 16   -0.411510833  22.564000336                 1
#> 17    0.252223448 -17.010244293                 1
#> 18   -0.891921127  -5.308315112                 1
#> 19    0.435683299 -15.430520672                 1
#> 20   -1.237538422  -3.788981080                 1
#> 21   -0.224267885   2.791272389                 1
#> 22    0.377395646  15.323135274                 1
#> 23    0.133336361  -3.990463239                 1
#> 24    0.804189510 -10.787373750                 1
#> 25   -0.057106774  -2.766428228                 1
#> 26    0.503607972   5.589487825                 1
#> 27    1.085769362   1.637119723                 1
#> 28   -0.690953840  11.865608381                 1
#> 29   -1.284599354  -2.690183821                 1
#> 30    0.046726172 -14.833995679                 1
#> 31   -0.235706556   8.563486227                 1
#> 32   -0.542888255 -10.018106578                 1
#> 33   -0.433310317   5.552698281                 1
#> 34   -0.649471647   4.621443448                 1
#> 35    0.726750747   3.325131934                 1
#> 36    1.151911754  15.515047175                 1
#> 37    0.992160365  -1.855301100                 1
#> 38   -0.429513109 -11.827756259                 1
#> 39    1.238304101   0.571266866                 2
#> 40   -0.279346282  -2.087797060                 1
#> 41    1.757903090   5.715542948                 1
#> 42    0.560746091  14.711798569                 1
#> 43   -0.452783973 -16.274948214                 1
#> 44   -0.832043296  -1.924973563                 1
#> 45   -1.166570547   0.633295272                 2
#> 46   -1.065590580   4.947369026                 1
#> 47   -1.563782051  -6.271411914                 1
#> 48    1.156536997   3.515093481                 1
#> 49    0.832047129   3.432701332                 1
#> 50   -0.227328691  -0.111354928                 2
#> 51    0.266137362  -1.119914878                 1
#> 52   -0.376702719  13.200701339                 1
#> 53    2.441364629   6.065358088                 1
#> 54   -0.795339117  17.607827369                 1
#> 55   -0.054877474  -2.421369111                 1
#> 56    0.250141323 -11.200963665                 1
#> 57    0.618243294  -3.561326303                 1
#> 58   -0.172623503   0.876569891                 1
#> 59   -2.223900274  -2.145482551                 2
#> 60   -1.263614385  -3.706451609                 1
#> 61    0.358728896 -14.194962440                 1
#> 62   -0.011045478  -7.175484858                 1
#> 63   -0.940649163  -9.866131045                 1
#> 64   -0.115825322   2.774778738                 1
#> 65   -0.814968709   3.103156113                 1
#> 66    0.242263481  13.856832788                 1
#> 67   -1.425098395  -2.965627731                 1
#> 68    0.365941123   8.535504157                 1
#> 69    0.248412649  -4.313300281                 1
#> 70    0.065288182 -13.633981155                 1
#> 71    0.019156392   3.910829903                 1
#> 72    0.257338377  -4.675830003                 1
#> 73   -0.649010078  -7.511031297                 1
#> 74   -0.119168762  -6.809828484                 1
#> 75    0.664135700   9.805531457                 1
#> 76    1.100969102  14.151896257                 1
#> 77    0.143771481   9.066279291                 1
#> 78   -0.117753598  -2.458421938                 1
#> 79   -0.912068367 -11.787115875                 1
#> 80   -1.437586241   2.003164478                 1
#> 81   -0.797089525  10.020898654                 1
#> 82    1.254083106   7.536549576                 1
#> 83    0.772142186   2.830650915                 1
#> 84   -0.219515627   1.999962649                 1
#> 85   -0.424810283  -7.592538509                 1
#> 86   -0.418980099   3.994247931                 1
#> 87    0.996986861   0.502173998                 2
#> 88   -0.275778029   0.611751506                 1
#> 89    1.256018817  -1.817655681                 1
#> 90    0.646674390 -10.422137286                 1
#> 91    1.299312303  -5.334689431                 1
#> 92   -0.873262112   6.894588456                 1
#> 93    0.008370960   0.350353050                 1
#> 94   -0.880871723   0.131821312                 2
#> 95    0.596259017  -1.676852712                 1
#> 96    0.119717641  12.605317446                 1
#> 97   -0.282173877   0.166370131                 2
#> 98    1.455988401   2.242764105                 1
#> 99    0.229019591   1.342986621                 1
#> 100   0.996543929  -8.669098624                 1
#> 101   0.781859185  -0.598200967                 2
#> 102  -0.776776622  11.582298370                 1
#> 103  -0.615989908   4.123127335                 1
#> 104   0.046580303 -13.068254611                 1
#> 105  -1.130385778   0.696080492                 2
#> 106   0.576718782   5.039057457                 1
#> 107  -1.280749432  -0.674519130                 2
#> 108   1.625447303   7.044279280                 1
#> 109  -0.500696596  -1.554017076                 1
#> 110   1.678297208  -9.461644006                 1
#> 111  -0.412519887   6.565061509                 1
#> 112  -0.972286836   2.363987184                 1
#> 113   0.025382868   4.892920732                 1
#> 114   0.027475337   9.369542685                 1
#> 115  -1.680182722   1.777555383                 1
#> 116   1.053750863 -14.666204586                 1
#> 117  -1.119599105   1.089362091                 1
#> 118   0.335617210 -14.736797569                 1
#> 119   0.494795767  -4.779387954                 1
#> 120   0.138052709   8.583118128                 1
#> 121  -0.118792026 -15.485855959                 1
#> 122   0.197684262   0.956885579                 1
#> 123  -1.068692711  -5.478016469                 1
#> 124  -0.803213217  -2.710772844                 1
#> 125  -1.113765136   8.785822722                 1
#> 126   1.580091684   4.104078024                 1
#> 127   1.497818761  11.649672607                 1
#> 128   0.262645459 -10.198819855                 1
#> 129  -1.232901200  -7.825143157                 1
#> 130  -0.003723534  -6.794732627                 1
#> 131   1.511672283  -1.166718152                 2
#> 132  -0.475698284  -9.016212019                 1
#> 133   0.797916438  -7.378814513                 1
#> 134  -0.974002561  -8.770972218                 1
#> 135   0.689372698   5.438784678                 1
#> 136  -0.955839103   4.939088294                 1
#> 137  -1.231707058   8.247893912                 1
#> 138  -0.956891881  23.954097300                 1
#> 139  -0.869782874  -1.622313589                 1
#> 140  -0.910680682   6.165132895                 1
#> 141   0.741276305  29.397730670                 1
#> 142   0.068511533   5.045404136                 1
#> 143  -0.323750755  -0.621155677                 1
#> 144  -1.086503047  -8.751986424                 1
#> 145  -1.015928947  -4.919279272                 1
#> 146  -0.767790185 -15.198230952                 1
#> 147  -1.119720061 -14.151354291                 1
#> 148  -0.448174237  -3.644884426                 1
#> 149   0.471736374   2.873577762                 1
#> 150  -1.180490683   0.363849135                 2
#> 151   1.470256997  -3.510086065                 1
#> 152  -1.311420592 -16.373000071                 1
#> 153  -0.096524923   5.932626392                 1
#> 154   2.369719908   4.136595054                 1
#> 155   0.890626476  14.549637031                 1
#> 156  -0.252183161 -16.705714453                 1
#> 157  -0.865763755  -2.581414949                 1
#> 158   0.582586000  15.752896998                 1
#> 159  -0.012529347   1.047722514                 1
#> 160  -0.374854762  12.458278424                 1
#> 161   0.317885735   5.167988214                 1
#> 162  -0.488805635   1.228417308                 1
#> 163   2.658658027   8.227943901                 1
#> 164   1.680278205 -16.207436862                 1
#> 165   0.779584009  -3.058925752                 1
#> 166   0.713240520   5.456381150                 1
#> 167  -0.542881937  12.070172795                 1
#> 168   0.885778374   6.905585591                 1
#> 169  -0.348594685   1.743531002                 1
#> 170  -1.008054578  10.265100215                 1
#> 171   1.883182542   0.124783240                 2
#> 172  -0.928971079  -9.947753191                 1
#> 173  -0.294196454  -0.226463752                 2
#> 174  -0.614950271  -1.473060039                 1
#> 175  -0.947075792   3.330537731                 1
#> 176   0.598975150  -3.427420825                 1
#> 177  -1.523614882   5.876571324                 1
#> 178  -0.206189002  18.552076255                 1
#> 179  -0.574295414 -16.169804432                 1
#> 180  -1.390166037   5.256694120                 1
#> 181  -0.070417383  -6.504778115                 1
#> 182  -0.430879530  -5.662481985                 1
#> 183  -0.592225373 -16.345854493                 1
#> 184   0.981116160  -2.333601885                 1
#> 185   0.532409357   3.011690461                 1
#> 186  -0.090456124 -12.844508514                 1
#> 187   0.156490492  17.447654098                 1
#> 188  -0.737311691  -6.835771919                 1
#> 189  -0.201341206 -20.508985274                 1
#> 190   1.102176595  -1.026456898                 2
#> 191  -0.016748256  21.166709991                 1
#> 192   0.161788634  14.366887229                 1
#> 193   2.024761390  11.497009599                 1
#> 194  -0.703694254   7.100737723                 1
#> 195   0.960792384   4.153188441                 1
#> 196   1.790485054  -3.942622204                 1
#> 197  -1.064165163 -13.570263083                 1
#> 198   0.017636546 -20.006515182                 1
#> 199  -0.389908629 -10.608051982                 1
#> 200  -0.490832752 -16.044842013                 1
#> 201  -1.045717652  -8.891272253                 1
#> 202  -0.896211264   6.556398195                 1
#> 203   1.269387164  -7.962164231                 1
#> 204   0.593840949 -13.845972754                 1
#> 205   0.775634319  -8.750041223                 1
#> 206   1.557370376 -15.463225713                 1
#> 207  -0.365401797   7.233817853                 1
#> 208   0.816556449 -13.516084530                 1
#> 209  -0.060634778  -1.309491640                 1
#> 210  -0.501378318   5.215127511                 1
#> 211   0.926062725  10.813730185                 1
#> 212   0.036937691  17.045197079                 1
#> 213  -1.066200174 -15.842037091                 1
#> 214  -0.238456353   8.320921224                 1
#> 215   1.495223444  -5.008879512                 1
#> 216   1.172158547  -1.625257244                 1
#> 217  -1.457707210  13.026962568                 1
#> 218   0.095056227  -5.464183076                 1
#> 219   0.847664964   6.114261873                 1
#> 220  -1.624364530  -0.842018763                 2
#> 221   1.408563357  -4.410776626                 1
#> 222  -0.541760362  12.695934450                 1
#> 223   0.278664724  -2.021164108                 1
#> 224  -0.193972745  -1.912459406                 1
#> 225   1.576158181   6.267406264                 1
#> 226  -1.475547635   8.236642547                 1
#> 227  -0.144608207  -8.310368843                 1
#> 228  -1.075010191  10.321859366                 1
#> 229   0.406542732  -5.722785388                 1
#> 230   2.229262202  -7.977989776                 1
#> 231  -1.514497008 -20.998230162                 1
#> 232  -0.061707422  -1.309417242                 1
#> 233  -0.147270790   2.852676214                 1
#> 234   1.541593069  -6.367229890                 1
#> 235  -0.981855669  11.179400023                 1
#> 236   0.496578173   5.581369757                 1
#> 237   1.696947881   0.899127608                 2
#> 238  -0.260736309  16.269331389                 1
#> 239  -0.705928586 -13.521869810                 1
#> 240  -0.161178506   2.570764187                 1
#> 241   0.501321828   7.611362679                 1
#> 242  -1.013539670  -8.958083738                 1
#> 243   1.614752235  -2.311687121                 1
#> 244   0.005641985  -0.502704390                 1
#> 245  -2.904899060  -4.005047159                 1
#> 246  -1.107164819   0.627255927                 2
#> 247   1.547566933  -1.392045354                 2
#> 248  -0.976830350  -7.481371135                 1
#> 249  -0.101503448   6.853899171                 1
#> 250   0.042650250  -5.188562393                 1
#> 251  -0.797552249  -0.626368228                 2
#> 252  10.322980923   0.481335326                 2
#> 253   0.781313160   1.695271084                 1
#> 254  -2.655045308  -1.761226294                 2
#> 255   4.610996449   0.198013015                 2
#> 256   2.549822711   0.397349099                 2
#> 257   3.460893446   0.029225495                 2
#> 258   9.043239051   2.560273389                 2
#> 259  -2.573692121   1.257127712                 2
#> 260  -2.136059016  -0.534537686                 2
#> 261  -1.833069305  -0.625227429                 2
#> 262   9.652248051   0.913848687                 2
#> 263  17.552500071   1.007199535                 2
#> 264   8.680086602   0.719291823                 2
#> 265   9.345641296  -0.604711661                 2
#> 266  15.919862434   0.539054406                 2
#> 267  -5.144851323  -0.076830886                 2
#> 268 -13.232546008   1.849919560                 2
#> 269  -9.993298001  -0.854907551                 2
#> 270   2.289630502   0.032637295                 2
#> 271   0.210102128  -1.025059481                 1
#> 272 -24.443327651  -0.982249076                 2
#> 273  10.667603263   0.004101957                 2
#> 274  -1.977796778  -0.233427178                 2
#> 275  -0.510054846  -0.498888219                 2
#> 276   3.659080585   1.549712963                 2
#> 277 -11.784869547   0.087496917                 2
#> 278  -6.360221582   1.318701131                 2
#> 279   9.305025029  -0.981224119                 2
#> 280  17.754848707  -0.245622588                 2
#> 281   6.187725661  -1.403933837                 2
#> 282   6.644182950   1.440893147                 2
#> 283 -15.374432086  -0.981359991                 2
#> 284   9.473578117   1.474244904                 2
#> 285  10.109151981  -0.991197245                 2
#> 286  -3.191328052  -0.094497345                 2
#> 287   1.694149680  -2.875141684                 1
#> 288   8.314200246  -0.246866101                 2
#> 289  19.374307588   0.014744490                 2
#> 290  -9.987654084  -1.919087703                 2
#> 291   9.265573999  -0.287813744                 2
#> 292  12.394392748  -0.346637445                 2
#> 293   8.233299467  -1.839588585                 2
#> 294   2.642610883   0.898588941                 2
#> 295  -1.365647622  -1.212855011                 2
#> 296   1.484022958  -0.218964232                 2
#> 297   7.817196875   0.564268161                 2
#> 298  -9.702310440  -0.525434376                 2
#> 299 -11.000987714   0.744374225                 2
#> 300  -6.403002264   0.128981753                 2
#> 301 -12.816284997   1.488274257                 2
#> 302 -15.024099707  -0.662681951                 2
#> 303  12.413125191  -1.160655001                 2
#> 304  -8.277071250   0.358774234                 2
#> 305  -4.540410394  -0.194846384                 2
#> 306 -10.212586314  -0.295282008                 2
#> 307   4.614957495   0.496640424                 2
#> 308   4.751899044   0.484912788                 2
#> 309  10.694225916   0.018784501                 2
#> 310  -4.158327218   0.634774565                 2
#> 311  11.169623593   0.754444088                 2
#> 312   1.233193967   0.833589035                 2
#> 313 -11.696807150   0.965761296                 2
#> 314  10.362680483   1.293879968                 2
#> 315  13.621027301  -0.136551021                 2
#> 316   7.219240884  -0.440138688                 2
#> 317  -3.874832476  -1.227283914                 2
#> 318 -12.102308462  -0.237653070                 2
#> 319   0.336978632  -0.926858157                 1
#> 320 -12.111496103   0.411235356                 2
#> 321 -16.326038843  -0.198864577                 2
#> 322   6.336301786  -0.557449639                 2
#> 323  -2.691176974  -0.977157122                 2
#> 324  -3.590021666   0.060735249                 2
#> 325  -5.789855969  -0.597229480                 2
#> 326  11.840445493  -1.258948697                 2
#> 327 -14.264710933  -1.410047202                 2
#> 328  -4.000467824   1.101304079                 2
#> 329   5.095647701  -0.677242026                 2
#> 330   2.612931070  -0.762173508                 2
#> 331  -5.539898991  -0.291748602                 2
#> 332  -0.558070369  -0.575188529                 2
#> 333  -1.733492295  -0.443941868                 2
#> 334  15.148385537  -0.312570444                 2
#> 335   0.923164615  -0.603004426                 2
#> 336  -5.893517821  -1.093934720                 2
#> 337  10.679176797   0.714706191                 2
#> 338 -11.451327046  -0.108812263                 2
#> 339   4.842581221  -1.443797935                 2
#> 340  11.077653524   0.806123264                 2
#> 341 -10.491362671  -1.739835078                 2
#> 342   2.917885035  -0.401320309                 2
#> 343   8.557806375  -0.287582489                 2
#> 344  10.056902417  -0.938407400                 2
#> 345  13.715721419   0.287667139                 2
#> 346  17.262435687  -1.505401124                 2
#> 347   1.801595752   1.519297013                 2
#> 348  -7.624714453   0.367409356                 2
#> 349  -6.811807858   1.699862409                 2
#> 350   6.255941547   0.644196977                 2
#> 351 -14.156260870  -1.687801029                 2
#> 352   4.502104199   0.647645994                 2
#> 353 -10.394267936   0.448794227                 2
#> 354  -8.326811055   1.026302158                 2
#> 355   3.077141028   1.074978223                 2
#> 356  -5.306651663   0.458309613                 2
#> 357  -7.455945433   0.631586758                 2
#> 358  -5.973571888  -0.580466398                 2
#> 359  12.747967674   1.584192142                 2
#> 360 -12.343521084  -1.763992940                 2
#> 361   0.979746508  -1.880621970                 1
#> 362  -4.201400270  -1.291719044                 2
#> 363  26.655690270   0.909670446                 2
#> 364  -1.420444621  -1.107755682                 2
#> 365  16.374827460  -0.384123875                 2
#> 366   7.773682959   0.082734834                 2
#> 367 -14.097988979  -0.483882474                 2
#> 368   4.639103719  -2.084741125                 2
#> 369   6.265975609   1.167586964                 2
#> 370   0.228789905  -0.076825771                 2
#> 371 -17.566049711   0.530421395                 2
#> 372  20.208228628   0.004908796                 2
#> 373  -1.689634579  -0.530231430                 2
#> 374 -12.611260707   0.102202410                 2
#> 375  11.235226552   0.815459636                 2
#> 376  -2.223704074  -1.506521863                 2
#> 377   2.210035640  -1.157180172                 2
#> 378  -7.008512017   1.301554937                 2
#> 379 -16.117975627  -0.905640775                 2
#> 380  -1.507714704   0.015563481                 2
#> 381   6.758819432  -1.428005197                 2
#> 382  -0.153210630   0.687979702                 1
#> 383   0.248210766   0.594196501                 1
#> 384   1.999315324  -0.270315244                 2
#> 385   3.587686423   1.554076076                 2
#> 386  -2.353665715  -0.510742458                 2
#> 387  -2.728598617  -0.291842736                 2
#> 388  -6.602146063   1.101438165                 2
#> 389 -12.436692423   0.458665215                 2
#> 390 -12.975186359   0.851908039                 2
#> 391  -7.076791478   1.466199919                 2
#> 392  -6.682340156   0.243863295                 2
#> 393  -3.518822223   0.649421934                 2
#> 394 -10.941795563  -0.641118156                 2
#> 395  -8.013650322   0.464459771                 2
#> 396  13.484453680  -0.309200447                 2
#> 397   3.346551999   1.276020709                 2
#> 398   2.349954751  -0.047611220                 2
#> 399  -0.220894711   0.030101166                 2
#> 400  -8.259203192  -0.003580270                 2
#> 401  -5.326957960  -2.034572643                 2
#> 402  -3.338937501  -0.992819963                 2
#> 403   0.791318315   1.963858056                 1
#> 404  -0.312537103   0.706528773                 1
#> 405  16.257368421   0.039674980                 2
#> 406  -3.062124587  -0.158518590                 2
#> 407  -6.740669300  -0.162257551                 2
#> 408 -14.752231548   0.438821912                 2
#> 409  -9.010689082   0.780699292                 2
#> 410  23.361946541  -0.981205622                 2
#> 411   2.728789336  -0.141585732                 2
#> 412   8.178726494  -1.256248953                 2
#> 413   1.870649530  -0.429033676                 2
#> 414   1.602126018   0.508776266                 2
#> 415  -1.491884984  -1.446889743                 2
#> 416   5.013932580   1.019512829                 2
#> 417  12.999009631   1.178546976                 2
#> 418   8.112213546  -0.010258765                 2
#> 419  -1.998315400   0.268624871                 2
#> 420   0.955721841   1.342028872                 1
#> 421 -13.042987501  -0.583606443                 2
#> 422  10.246965512  -0.937000891                 2
#> 423  16.324422397  -1.152157056                 2
#> 424 -13.599348594  -0.975183431                 2
#> 425  -0.162781728  -0.711413598                 1
#> 426  -7.921692885   0.371912476                 2
#> 427 -10.777293894  -0.943514417                 2
#> 428   9.618830620  -0.274367308                 2
#> 429  10.500631726   0.154468494                 2
#> 430  18.269626463  -1.288653202                 2
#> 431   4.501554668   1.419102319                 2
#> 432 -16.404748009   1.307827762                 2
#> 433   4.402449548  -1.804975774                 2
#> 434  -6.245004574  -0.484069527                 2
#> 435  -1.791580607  -0.373211780                 2
#> 436   2.177586688   0.169085902                 2
#> 437   0.633172530  -1.172983603                 1
#> 438   7.147314594  -0.376196737                 2
#> 439  -5.898408265   0.783127640                 2
#> 440  -6.097185874   0.991171304                 2
#> 441   1.654366629   0.229284475                 2
#> 442   7.899374403   0.336850949                 2
#> 443  -3.530078214  -0.647231999                 2
#> 444  -6.020534746  -0.443206682                 2
#> 445  -9.091199188   1.161751761                 2
#> 446  -6.757475362   0.065051010                 2
#> 447  -2.551976970  -0.682467580                 2
#> 448 -10.492226833   1.634341658                 2
#> 449   6.221024872  -0.584849244                 2
#> 450 -18.508838677   1.129936001                 2
#> 451  -1.387570525   0.581201043                 2
#> 452 -13.004048795   0.379293067                 2
#> 453  10.945357773  -0.310740876                 2
#> 454  -4.600540233   0.886390001                 2
#> 455  10.156848815  -1.641864750                 2
#> 456  -3.776789414  -0.988563742                 2
#> 457  -1.914657611  -0.244003429                 2
#> 458  -9.185947020   0.156056926                 2
#> 459 -27.741272559   0.102051958                 2
#> 460  14.815290267  -0.287458172                 2
#> 461 -12.700926056  -0.293194886                 2
#> 462  23.086238678   0.469477058                 2
#> 463   2.930343425  -0.662588512                 2
#> 464  -5.089840368  -2.190372666                 2
#> 465   6.290088088   0.003854672                 2
#> 466   7.926435876   0.862942829                 2
#> 467  -2.049794739   0.980223359                 2
#> 468   1.226706740  -0.291724771                 2
#> 469  16.590674096  -0.061013476                 2
#> 470 -12.648525462   1.511868401                 2
#> 471  -0.441747634  -0.643609065                 1
#> 472   7.147390560  -0.130950479                 2
#> 473 -12.726211639   0.485211453                 2
#> 474 -16.109346268  -0.645182435                 2
#> 475  12.189575992  -0.463956220                 2
#> 476  -6.729900347   0.567249697                 2
#> 477  -9.337072301  -0.723407834                 2
#> 478  -0.325549223   0.455884554                 1
#> 479  -4.989155458  -1.197934626                 2
#> 480  -5.029800765  -2.018180550                 2
#> 481 -12.778860448   0.619723405                 2
#> 482  -8.171638687  -0.132857466                 2
#> 483   7.628697851  -0.434960740                 2
#> 484   4.495167897  -0.521754179                 2
#> 485  -5.740429658   0.992288426                 2
#> 486  21.739546330  -1.085422408                 2
#> 487  13.328173125   0.959832414                 2
#> 488  -6.650835828  -1.039010541                 2
#> 489  -0.879886196  -0.137841967                 2
#> 490 -29.127471591  -0.214708418                 2
#> 491   8.613572923   0.573718585                 2
#> 492   5.034007572  -1.776449315                 2
#> 493  -4.479306416  -0.256392295                 2
#> 494  -1.745822786  -0.138550636                 2
#> 495   0.777014412   0.557684974                 2
#> 496   9.157567844   1.122346289                 2
#> 497  -8.729262206  -0.982147207                 2
#> 498   9.597171614   0.326905556                 2
#> 499   6.799582393   0.447826608                 2
#> 500  -6.329270648   1.183724120                 2