Classifies Data According to Unique Observations

Gives a one-to-one mapping from unique observations to rows of a data matrix.

partuniq(x)

Arguments

x
Matrix of observations.

Value

A vector of length nrow(x) with integer entries. An observation k is assigned an integer i whenever observation i is the first row of x that is identical to observation k (note that i <= k).

See also

Examples

set.seed(0)

mat <- data.frame(lets = sample(LETTERS[1:2],9,TRUE), nums = sample(1:2,9,TRUE))
mat

#>   lets nums
#> 1    B    2
#> 2    A    1
#> 3    A    1
#> 4    B    1
#> 5    B    2
#> 6    A    1
#> 7    B    2
#> 8    B    1
#> 9    B    2

ans <- partuniq(mat)
ans

#> [1] 1 2 2 4 1 2 1 4 1

partconv(ans,consec=TRUE)

#> [1] 1 2 2 3 1 2 1 3 1