Summary Function for Bootstrap Inference for Gaussian Finite Mixture Models

Summary of bootstrap distribution for the parameters of a Gaussian mixture model providing either standard errors or percentile bootstrap confidence intervals.

Usage

# S3 method for MclustBootstrap
summary(object, what = c("se", "ci", "ave"), conf.level = 0.95, ...)

Arguments

object

An object of class 'MclustBootstrap' as returned by MclustBootstrap.

what

A character string: "se" for the standard errors; "ci" for the confidence intervals; "ave" for the averages.

conf.level

A value specifying the confidence level of the interval.

...

Further arguments passed to or from other methods.

Details

For details about the procedure used to obtain the bootstrap distribution see MclustBootstrap.

See also

Examples

# \donttest{
data(diabetes)
X = diabetes[,-1]
modClust = Mclust(X) 
bootClust = MclustBootstrap(modClust)
summary(bootClust, what = "se")
#> ---------------------------------------------------------- 
#> Resampling standard errors 
#> ---------------------------------------------------------- 
#> Model                      = VVV 
#> Num. of mixture components = 3 
#> Replications               = 999 
#> Type                       = nonparametric bootstrap 
#> 
#> Mixing probabilities:
#>          1          2          3 
#> 0.05122683 0.05182469 0.03558267 
#> 
#> Means:
#>                1         2        3
#> glucose 1.034526  3.421991 16.44573
#> insulin 7.573130 29.138785 64.52080
#> sspg    7.790168 29.851418 10.12022
#> 
#> Variances:
#> [,,1]
#>          glucose   insulin      sspg
#> glucose 11.60153  53.36198  51.64031
#> insulin 53.36198 498.28283 403.97374
#> sspg    51.64031 403.97374 612.00879
#> [,,2]
#>           glucose   insulin      sspg
#> glucose  62.33107  602.2394  431.8649
#> insulin 602.23945 7319.3483 3155.4451
#> sspg    431.86487 3155.4451 6677.0051
#> [,,3]
#>           glucose   insulin      sspg
#> glucose  981.2215  4087.832  646.5159
#> insulin 4087.8323 19042.861 2505.3844
#> sspg     646.5159  2505.384  501.4012
summary(bootClust, what = "ci")
#> ---------------------------------------------------------- 
#> Resampling confidence intervals 
#> ---------------------------------------------------------- 
#> Model                      = VVV 
#> Num. of mixture components = 3 
#> Replications               = 999 
#> Type                       = nonparametric bootstrap 
#> Confidence level           = 0.95 
#> 
#> Mixing probabilities:
#>               1         2         3
#> 2.5%  0.4469619 0.1495901 0.1325502
#> 97.5% 0.6439540 0.3610308 0.2690097
#> 
#> Means:
#> [,,1]
#>        glucose  insulin     sspg
#> 2.5%  89.12748 343.8486 150.5996
#> 97.5% 93.04433 374.2694 181.0273
#> [,,2]
#>         glucose  insulin     sspg
#> 2.5%   98.77345 451.3980 261.5801
#> 97.5% 111.95854 561.0686 376.9218
#> [,,3]
#>        glucose   insulin      sspg
#> 2.5%  197.2114  968.2317  62.82878
#> 97.5% 261.3064 1220.3246 101.77019
#> 
#> Variances:
#> [,,1]
#>        glucose  insulin     sspg
#> 2.5%  37.39976 1266.724 1513.569
#> 97.5% 83.23178 3177.774 3904.742
#> [,,2]
#>         glucose   insulin     sspg
#> 2.5%   84.53327  3242.897 12266.86
#> 97.5% 337.36342 31009.184 38525.69
#> [,,3]
#>        glucose   insulin     sspg
#> 2.5%  3386.800  47497.23 1343.209
#> 97.5% 7226.332 120173.04 3312.181

data(acidity)
modDens = densityMclust(acidity, plot = FALSE)
modDens = MclustBootstrap(modDens)
summary(modDens, what = "se")
#> ---------------------------------------------------------- 
#> Resampling standard errors 
#> ---------------------------------------------------------- 
#> Model                      = E 
#> Num. of mixture components = 2 
#> Replications               = 999 
#> Type                       = nonparametric bootstrap 
#> 
#> Mixing probabilities:
#>          1          2 
#> 0.03925417 0.03925417 
#> 
#> Means:
#>          1          2 
#> 0.04610371 0.06701070 
#> 
#> Variances:
#>          1          2 
#> 0.02317383 0.02317383 
summary(modDens, what = "ci")
#> ---------------------------------------------------------- 
#> Resampling confidence intervals 
#> ---------------------------------------------------------- 
#> Model                      = E 
#> Num. of mixture components = 2 
#> Replications               = 999 
#> Type                       = nonparametric bootstrap 
#> Confidence level           = 0.95 
#> 
#> Mixing probabilities:
#>               1         2
#> 2.5%  0.5445248 0.3048908
#> 97.5% 0.6951092 0.4554752
#> 
#> Means:
#>              1        2
#> 2.5%  4.281097 6.186432
#> 97.5% 4.458207 6.442210
#> 
#> Variances:
#>               1         2
#> 2.5%  0.1393808 0.1393808
#> 97.5% 0.2311163 0.2311163
# }